poistamisominaisuuden parantaminen

+++ application/reviews/views.py
@@ -33,11 +33,12 @@ def reviews_create():
 @login_required
 def review_delete(review_id):
     review = Review.query.get(review_id)
-    #if book.account_id != current_user.id:
-        #lopullinen tekeminen päättämättä
-       # return login_manager.unauthorized()
+    book_id = review.book_id
+    if review.account_id != current_user.id:
+        #Ehkä jokin viesti siitä, ettei arvostelua voi poistaa, koska joku muu on sen kirjoittanut tai edes paluu samaa näkymää 
+        return redirect(url_for("book_view", book_id = book_id))
     
     db.session.delete(review)
     db.session.commit()
 
-    return redirect(url_for("books_index"))
+    return redirect(url_for("book_view", book_id = book_id))
